+++ CHANGELOG.md
@@ -12,6 +12,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - New `FireManager` interface has a void `FireManager#startFires()` method
 - `DefaultFireManager` updated to implement `FireManager`
+- `FlammabilityChecker` and default implementation `DefaultFlammabilityChecker`
+  that indicate whether a specific grid cell is flammable or not
 
 #### CHANGED